refactor: use sudo command to remove file
refactor: code organization
