+++ cmd/eibc/fulfill/denoms/remove/remove.go
@@ -5,13 +5,12 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/dymensionxyz/roller/cmd/consts"
+	"github.com/dymensionxyz/roller/utils/eibc"
+	"github.com/dymensionxyz/roller/utils/filesystem"
 	"github.com/pterm/pterm"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v3"
-
-	"github.com/dymensionxyz/roller/cmd/consts"
-	globalutils "github.com/dymensionxyz/roller/utils"
-	"github.com/dymensionxyz/roller/utils/eibc"
 )
 
 func Cmd() *cobra.Command {
@@ -27,7 +26,7 @@ func Cmd() *cobra.Command {
 			}
 
 			eibcHome := filepath.Join(home, consts.ConfigDirName.Eibc)
-			isEibcClientInitialized, err := globalutils.DirNotEmpty(eibcHome)
+			isEibcClientInitialized, err := filesystem.DirNotEmpty(eibcHome)
 			if err != nil {
 				pterm.Error.Println("failed to check eibc client initialized", err)
 				return
